+++ malachite-q/src/bin_util/demo_and_bench/rational_polynomial/comparison/mod.rs
@@ -6,10 +6,12 @@
 // Lesser General Public License (LGPL) as published by the Free Software Foundation; either version
 // 3 of the License, or (at your option) any later version. See <https://www.gnu.org/licenses/>.
 
-pub mod bench;
-pub mod common;
-pub mod extra_variadic;
-pub mod gaussian_rational;
-pub mod generators;
-pub mod rational;
-pub mod rational_polynomial;
+use malachite_base::test_util::runner::Runner;
+
+pub(crate) fn register(runner: &mut Runner) {
+    cmp::register(runner);
+    shortlex_cmp::register(runner);
+}
+
+mod cmp;
+mod shortlex_cmp;