+++ malachite-q/src/bin_util/demo_and_bench/rational_polynomial/comparison/shortlex_cmp.rs
@@ -1,0 +1,61 @@
+// Copyright © 2026 Mikhail Hogrefe
+//
+// This file is part of Malachite.
+//
+// Malachite is free software: you can redistribute it and/or modify it under the terms of the GNU
+// Lesser General Public License (LGPL) as published by the Free Software Foundation; either version
+// 3 of the License, or (at your option) any later version. See <https://www.gnu.org/licenses/>.
+
+use malachite_base::test_util::bench::{BenchmarkType, run_benchmark};
+use malachite_base::test_util::generators::common::{GenConfig, GenMode};
+use malachite_base::test_util::runner::Runner;
+use malachite_q::rational_polynomial::ShortlexRationalPolynomialRef;
+use malachite_q::test_util::bench::bucketers::pair_rational_polynomial_max_bit_bucketer;
+use malachite_q::test_util::generators::rational_polynomial_pair_gen;
+use malachite_q::test_util::rational_polynomial::comparison::cmp::*;
+
+pub(crate) fn register(runner: &mut Runner) {
+    register_demo!(runner, demo_shortlex_rational_polynomial_cmp);
+
+    register_bench!(
+        runner,
+        benchmark_shortlex_rational_polynomial_cmp_algorithms
+    );
+}
+
+fn demo_shortlex_rational_polynomial_cmp(gm: GenMode, config: &GenConfig, limit: usize) {
+    for (p, q) in rational_polynomial_pair_gen().get(gm, config).take(limit) {
+        println!(
+            "({}).cmp(&{}) = {:?}",
+            p,
+            q,
+            ShortlexRationalPolynomialRef(&p).cmp(&ShortlexRationalPolynomialRef(&q))
+        );
+    }
+}
+
+#[allow(unused_must_use)]
+fn benchmark_shortlex_rational_polynomial_cmp_algorithms(
+    gm: GenMode,
+    config: &GenConfig,
+    limit: usize,
+    file_name: &str,
+) {
+    run_benchmark(
+        "ShortlexRationalPolynomialRef.cmp(&ShortlexRationalPolynomialRef)",
+        BenchmarkType::Algorithms,
+        rational_polynomial_pair_gen().get(gm, config),
+        gm.name(),
+        limit,
+        file_name,
+        &pair_rational_polynomial_max_bit_bucketer("p", "q"),
+        &mut [
+            ("default", &mut |(p, q)| {
+                no_out!(ShortlexRationalPolynomialRef(&p).cmp(&ShortlexRationalPolynomialRef(&q)));
+            }),
+            ("materializing each coefficient", &mut |(p, q)| {
+                no_out!(rational_polynomial_shortlex_cmp_naive(&p, &q));
+            }),
+        ],
+    );
+}