+++ malachite-q/src/bin_util/demo_and_bench/rational_polynomial/mod.rs
@@ -10,8 +10,10 @@ use malachite_base::test_util::runner::Runner;
 
 pub(crate) fn register(runner: &mut Runner) {
     basic::register(runner);
+    comparison::register(runner);
     conversion::register(runner);
 }
 
 mod basic;
+mod comparison;
 mod conversion;